tools/logdump: custom cmd decoder
